refactor(gateway): isolate session-specific model metadata
Keep the uncached synchronous session ownership overlay separate from the
shared metadata cache and lifecycle owner. Preserve projection/config/read
ordering, missing-row behavior and native/host auth distinctions.

The exact CI merge with main's stop/activeWork handling exceeded max-lines.
This extraction reduces it from 728 to 695 logical lines without suppressing
the guard or changing lifecycle behavior. No stored representation or public
config changes. Net production growth is four module/type boundary lines.

Proof: exact merged-file lint; 80 metadata/shutdown/lifecycle tests; changed
checks; normal build; fresh Codex P0 autoreview. Follow-up for #136754.
